// src/modules/schools/infrastructure/types/schools-table-row.ts
export type SchoolsTableRow = {
  School_Id: string;
  Org_Name: string;
  Telephone: string;
  Fax: string;
  Email: string;
  Contact1_Name: string;
  URL: string;
  Add1_Line1: string;
  Add1_Suburb: string;
  Add1_City: string;
  Add2_Line1: string;
  Add2_Suburb: string;
  Add2_City: string;
  Add2_Postal_Code: string;
  Urban_Rural_Indicator: string;
  Org_Type: string;
  Definition: string;
  Authority: string;
  School_Donations: string;
  CoEd_Status: string;
  KMEPeakBody: string;
  Takiwā: string;
  Territorial_Authority: string;
  Regional_Council: string;
  Local_Office_Name: string;
  Education_Region: string;
  General_Electorate: string;
  Māori_Electorate: string;
  Statistical_Area_2_Code: string;
  Statistical_Area_2_Description: string;
  Ward: string;
  Col_Id: string;
  Col_Name: string;
  Latitude: string; // number
  Longitude: string; // number
  Enrolment_Scheme: string;
  Roll_Date: string; // timestamp
  Total: string; // number
  European: string; // number
  Māori: string; // number
  Pacific: string; // number
  Asian: string; // number
  MELAA: string; // number
  Other: string; // number
  International: string; // number
  Isolation_Index: string;
  Language_of_Instruction: string;
  BoardingFacilities: string;
  CohortEntry: string;
  Status: string;
  DateSchoolOpened: string; // timestamp
};